Move the fork assertion to seperate file
